fix: import database process
