+++ apps/gauzy/src/app/@shared/time-tracker/time-tracker.service.ts
@@ -9,7 +9,7 @@ import {
 	TimerState,
 	TimeLogSourceEnum
 } from '@gauzy/contracts';
-import { toLocal } from '@gauzy/common';
+import { toLocal } from '@gauzy/common-angular';
 import * as moment from 'moment';
 import { StoreConfig, Store, Query } from '@datorama/akita';
 import { Store as AppStore } from '../../@core/services/store.service';
