+++ apps/gauzy/src/app/pages/employees/activity/app-url-activity/app-url-activity/app-url-activity.component.ts
@@ -11,7 +11,7 @@ import {
 	IActivity
 } from '@gauzy/contracts';
 import { debounceTime } from 'rxjs/operators';
-import { toUTC, toLocal } from '@gauzy/common';
+import { toUTC, toLocal } from '@gauzy/common-angular';
 import { ActivityService } from 'apps/gauzy/src/app/@shared/timesheet/activity.service';
 import { ActivatedRoute } from '@angular/router';
 import * as _ from 'underscore';
