+++ apps/gauzy/src/app/pages/employees/activity/screenshot/screenshot/screenshot.component.ts
@@ -10,7 +10,7 @@ import { TimesheetService } from 'apps/gauzy/src/app/@shared/timesheet/timesheet
 import { debounceTime, filter } from 'rxjs/operators';
 import { Store } from 'apps/gauzy/src/app/@core/services/store.service';
 import { Subject } from 'rxjs';
-import { toUTC, toLocal } from '@gauzy/common';
+import { toUTC, toLocal } from '@gauzy/common-angular';
 import * as _ from 'underscore';
 import * as moment from 'moment';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
