+++ apps/gauzy/src/app/pages/employees/appointment/appointment.component.ts
@@ -36,8 +36,9 @@ import { AppointmentEmployeesService } from '../../../@core/services/appointment
 import { TimezoneSelectorComponent } from './timezone-selector/timezone-selector.component';
 import { TimeOffService } from '../../../@core/services/time-off.service';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { convertLocalToTimezone } from '@gauzy/common';
+import { convertLocalToTimezone } from '@gauzy/common-angular';
 import { ToastrService } from '../../../@core/services/toastr.service';
+
 @UntilDestroy({ checkProperties: true })
 @Component({
 	selector: 'ga-appointment-calendar',
