+++ apps/gauzy/src/app/pages/inventory/components/table-components/item-img-tags-row.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { getContrastColor } from '@gauzy/common';
+import { getContrastColor } from '@gauzy/common-angular';
 import { ViewCell } from 'ng2-smart-table';
 import { ComponentLayoutStyleEnum } from '@gauzy/contracts';
